+++ src/components/common/functionButton.module.css
@@ -5,8 +5,8 @@
 
     width: auto;
     min-width: 56px;
+    height: auto;
     min-height: 56px;
-    padding: 18px;
 
     border-radius: 12px;
