+++ src/components/memo/memoItem.module.css
